[IMP] ni_community_care: update filter my_area_employees
