+++ ni_community_care/views/ni_service_event_approval_views.xml
@@ -12,7 +12,7 @@
                 <filter
                     string="พนักงานในพื้นที่รับผิดชอบ"
                     name="my_area_employees"
-                    domain="[('employee_id.my_user_city_match','=',True)]"
+                    domain="[('user_id.my_user_city_match','=',True)]"
                     groups="ni_patient.group_manager"
                 />
                 <separator />
